% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{standard deviation}

\item{x}{input data}

\item{fun}{desired function to be executed}

\item{alpha}{percent interval estimate}

\item{cx}{cx}
}
\value{
bootstrap intervals
}
\description{
myboot2
}
\examples{
myboot2(10000,x=sam,fun=function(x) IQR(x),alpha=0.20,xlab="IQR(x)",col="purple",cx=1.5)
}
