% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt2.R
\name{myclt2}
\alias{myclt2}
\title{myclt}
\usage{
myclt2(n, iter)
}
\arguments{
\item{n}{mean}

\item{iter}{standard deviation}
}
\value{
histogram of the distribution
}
\description{
myclt
}
\examples{
m=myclt2(n=10,iter=10000)
}
