% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{lfun}

\item{x}{used to create plot, locate point of intersect with param}

\item{param}{used to create plot, locate point of intersect with x}
}
\value{
p vs y plot of liklihood
}
\description{
mymaxlik
}
\examples{
mymaxlik(x=c(3,3,4,3,4,5,5,4),param=seq(0,1,length=1000)
}
