% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{shaded region limit}
}
\value{
curve of the normal distribution
}
\description{
Myncurve
}
\examples{
x = myncurve(10,5,2)
}
