% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{mypvalue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{t0}

\item{xmax}{xmax}

\item{n}{n}

\item{alpha}{alpha}
}
\value{
plot and p-values
}
\description{
mypvalue
}
\examples{
myci(x = rnorm(30,mean=10,sd=12))
}
