% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{myread}
\usage{
myread(dird, csv)
}
\arguments{
\item{dird}{Directory of csv file to read in}

\item{csv}{name of csv file to read in}
}
\value{
a read-in csv file
}
\description{
myread
}
\examples{
x = myread("C:\\\\Users\\\\Austin\\\\Google Drive\\\\School\\\\Statistics\\\\Lab 2\\\\","EPAGAS.csv")
}
